OTel.API.Baggage.Core.Internal

Disclaimer

In general, changes to this module will not be reflected in the library's version updates. Direct use of this module should be done with utmost care, otherwise invariants will easily be violated.

class Monad m => MonadBaggage m where

Minimal complete definition

Nothing

Methods

getBaggage :: m Baggage

default getBaggage :: (MonadTrans t, MonadBaggage n, m ~ t n) => m Baggage

setBaggage :: Baggage -> m a -> m a

default setBaggage :: (MonadTransControl t, MonadBaggage n, m ~ t n) => Baggage -> m a -> m a

newtype Baggage

Constructors

Baggage

Fields

unBaggage :: HashMap Text Text

nullBaggage :: Baggage -> Bool

sizeBaggage :: Baggage -> Int

memberBaggage :: Key Text -> Baggage -> Bool

lookupBaggage :: Key Text -> Baggage -> Maybe Text

findWithDefaultBaggage :: Text -> Key Text -> Baggage -> Text

deleteBaggage :: Key Text -> Baggage -> Baggage

filterBaggage :: (Text -> Bool) -> Baggage -> Baggage

filterWithKeyBaggage :: (Key Text -> Text -> Bool) -> Baggage -> Baggage

foldMapWithKeyBaggage :: forall m. Monoid m => (Key Text -> Text -> m) -> Baggage -> m

toListBaggage :: Baggage -> [(Key Text, Text)]

newtype BaggageBuilder a

Constructors

BaggageBuilder

Fields

unBaggageBuilder :: Either (DList BaggageError) a

buildBaggage :: forall m. MonadThrow m => BaggageBuilder Baggage -> m Baggage

buildBaggagePure :: BaggageBuilder Baggage -> Either BaggageErrors Baggage

contextBackendBaggage :: ContextBackend Baggage

contextKeyBaggage :: ContextKey Baggage

isRFC7230TokenChar :: Char -> Bool

isRFC7230VCHARChar :: Char -> Bool